requirements.txt
fastapi
uvicorn
supabase
google-generativeai>=0.8.3
groq
python-multipart
apscheduler
requests
python-dotenv
pillow
